+++ src/main/java/com/diffplug/gradle/eclipserunner/JarFolderRunner.java
@@ -20,6 +20,8 @@ import com.diffplug.gradle.JRE;
 import java.io.File;
 import java.net.URL;
 import java.net.URLClassLoader;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -35,23 +37,37 @@ public class JarFolderRunner implements EclipseRunner {
 
 	@Override
 	public void run(List<String> args) throws Exception {
-		ClassLoader parent = null;
-		URL[] bootpath = null;
-		if (JRE.majorVersion() >= 9) {
-			// In J9+ the SystemClassLoader is a AppClassLoader. Thus we need it's parent
-			ClassLoader appClassLoader = ClassLoader.getSystemClassLoader();
-			parent = appClassLoader.getParent();
-			bootpath = JRE.getClasspath(appClassLoader);
-		} else {
-			// Running on Java 8
-			parent = ClassLoader.getSystemClassLoader();
-			bootpath = JRE.getClasspath(parent);
+		File plugins = new File(rootDirectory, "plugins");
+		List<URL> osgiClasspath = new ArrayList<>();
+		for (File plugin : plugins.listFiles()) {
+			if (plugin.isFile() && plugin.getName().endsWith(".jar")) {
+				osgiClasspath.add(plugin.toURI().toURL());
+			}
 		}
-		try (URLClassLoader classLoader = new URLClassLoader(bootpath, parent)) {
+		try (URLClassLoader classLoader = open(osgiClasspath)) {
 			Class<?> launcherClazz = classLoader.loadClass("com.diffplug.gradle.eclipserunner.EquinoxLauncher");
 			Object launcher = launcherClazz.getConstructor(File.class).newInstance(rootDirectory);
 			launcherClazz.getDeclaredMethod("setArgs", List.class).invoke(launcher, args);
 			launcherClazz.getDeclaredMethod("run").invoke(launcher);
 		}
+	}
+
+	public static URLClassLoader open(List<URL> urls) throws Exception {
+		ClassLoader parent = null;
+		URL[] boot;
+		if (JRE.majorVersion() >= 9) {
+			// In J9+ the SystemClassLoader is a AppClassLoader. Thus we need it's parent
+			ClassLoader appClassLoader = ClassLoader.getSystemClassLoader();
+			parent = appClassLoader.getParent();
+			boot = JRE.getClasspath(appClassLoader);
+		} else {
+			// Running on Java 8
+			parent = ClassLoader.getSystemClassLoader();
+			boot = JRE.getClasspath(parent);
+		}
+		List<URL> classpath = new ArrayList<>();
+		classpath.addAll(urls);
+		classpath.addAll(Arrays.asList(boot));
+		return new URLClassLoader(classpath.toArray(new URL[0]), parent);
 	}
 }