+++ src/main/java/com/diffplug/gradle/oomph/OomphIdeExtension.java
@@ -30,6 +30,7 @@ import com.diffplug.gradle.JavaExecable;
 import com.diffplug.gradle.Lazyable;
 import com.diffplug.gradle.StateBuilder;
 import com.diffplug.gradle.eclipserunner.EclipseIni;
+import com.diffplug.gradle.eclipserunner.EclipseIniLauncher;
 import com.diffplug.gradle.oomph.thirdparty.ConventionThirdParty;
 import com.diffplug.gradle.p2.P2Declarative;
 import com.diffplug.gradle.p2.P2Model;
@@ -56,6 +57,7 @@ import org.gradle.api.GradleException;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
 import org.gradle.api.XmlProvider;
+import org.gradle.api.file.FileCollection;
 import org.gradle.api.initialization.IncludedBuild;
 import org.gradle.api.specs.Spec;
 import org.gradle.api.specs.Specs;
@@ -470,7 +472,7 @@ public class OomphIdeExtension implements P2Declarative {
 	}
 
 	/** Sets up an IDE as described in this model from scratch. */
-	void ideSetupWorkspace() throws Exception {
+	void ideSetupWorkspace() throws Throwable {
 		if (workspaceExists()) {
 			project.getLogger().lifecycle("Skipping " + OomphIdePlugin.IDE_SETUP_WORKSPACE + " because it already exists, run " + OomphIdePlugin.IDE_CLEAN + " to force a rebuild.");
 		}
@@ -521,7 +523,7 @@ public class OomphIdeExtension implements P2Declarative {
 	}
 
 	/** Performs setup actions with a running OSGi container. */
-	private void internalSetup(File ideDir) throws IOException {
+	private void internalSetup(File ideDir) throws Throwable {
 		// get the user setup actions
 		List<SetupAction> list = setupActions.getResult();
 		// add the project importer
@@ -532,7 +534,14 @@ public class OomphIdeExtension implements P2Declarative {
 		ordered.add(new SaveWorkspace());
 
 		SetupWithinEclipse internal = new SetupWithinEclipse(ideDir, ordered);
-		Errors.constrainTo(IOException.class).run(() -> JavaExecable.exec(project, internal));
+		JavaExecable.exec(project, internal, execSpec -> {
+			FileCollection cp = execSpec.getClasspath().filter(file -> {
+				String name = file.getName();
+				return !name.startsWith("org.eclipse.") && !name.startsWith("biz.aQute.bndlib-");
+			})
+					.plus(project.files(EclipseIniLauncher.parseBundlesDotInfo(ideDir)));
+			execSpec.setClasspath(cp);
+		});
 	}
 
 	/////////